add configuration for pulsar/rabbitmq on alicloud
You may benchmark pulsar and rabbitmq on alicloud.
Some minor different when deploying with aws.
To avoid failure when downloading release package of pulsar/erlang/rabbitmq,
we will download them to localhost, and then upload them to ECS

In case of rabbitmq, only 1 client was used in the configuration of aws,
we use 4 to match other product configuration.