New max sustainable rate implementation (#329)
## Motivation

Currently the WorkloadGenerator.findMaximumSustainableRate algorithm introduces
publish delay. It has been observed the algorithm settles on a publish rate that
neither producers nor consumers can keep up with. For example, on a Kafka
workload with 100 topics, 1kb message, the algorithm settled on a publish rate
of 2.5m msg/s when the producers could only actually achieve 2m msg/s.

## Changes

Implement a new algorithm that checks for both receive backlog and publish
backlog and adjusts the publish according and has a progressive rate ramp up
when there is no observed backlog.